requirements.txt
wheel==0.45.1
langchain==1.1.0
langchain-openai==1.1.0
langchain-community==0.4.1
langchain-chroma==1.0.0
langchain-text-splitters==1.0.0
chromadb==1.3.5
pypdf==6.4.0
python-dotenv==1.2.1
chainlit==2.9.2
pytest==9.0.1
pytest-mock==3.15.1